requirements.txt
Django
djangorestframework
openai
python-dotenv
boto3
django-cors-headers
python-magic
Pillow
pytesseract
spacy
django-storages
gunicorn
django-filter
drf-spectacular
